@@ -55,15 +53,15 @@
 	<properties>
 		<!--		<license.verbose>true</license.verbose>-->
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.release>17</maven.compiler.release>
+		<maven.compiler.release>21</maven.compiler.release>
 		<maven.version>3.9.1</maven.version>
-		<maven-javadoc-plugin.version>3.5.0</maven-javadoc-plugin.version>
-		<guava.version>32.0.1-jre</guava.version>
-		<slf4j.version>2.0.7</slf4j.version>
-		<logback.version>1.4.8</logback.version>
-		<testng.version>7.8.0</testng.version>
-		<pouch.version>3.0</pouch.version>
-		<jmh.version>1.36</jmh.version>
+		<maven-javadoc-plugin.version>3.6.3</maven-javadoc-plugin.version>
+		<guava.version>33.0.0-jre</guava.version>
+		<slf4j.version>2.0.11</slf4j.version>
+		<logback.version>1.4.14</logback.version>
+		<testng.version>7.9.0</testng.version>
+		<pouch.version>4.0</pouch.version>
+		<jmh.version>1.37</jmh.version>
 		<pmd.version>6.55.0</pmd.version>
 	</properties>
 
