@@ -4,17 +4,14 @@
  */
 package com.github.cowwoc.requirements.test.java;
 
-import com.github.cowwoc.requirements.Requirements;
-import com.github.cowwoc.requirements.java.ValidationFailure;
 import com.github.cowwoc.requirements.java.internal.scope.ApplicationScope;
-import com.github.cowwoc.requirements.test.natives.internal.util.scope.TestApplicationScope;
+import com.github.cowwoc.requirements.test.TestValidatorsImpl;
+import com.github.cowwoc.requirements.test.scope.TestApplicationScope;
 import org.testng.annotations.Test;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import static com.github.cowwoc.requirements.natives.terminal.TerminalEncoding.NONE;
+import static com.github.cowwoc.requirements.java.terminal.TerminalEncoding.NONE;
 
 public final class BooleanTest
 {
@@ -24,7 +21,7 @@ public final class BooleanTest
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			boolean actual = false;
-			new Requirements(scope).requireThat(actual, "actual").isTrue();
+			new TestValidatorsImpl(scope).requireThat(actual, "Actual").isTrue();
 		}
 	}
 
@@ -34,7 +31,7 @@ public final class BooleanTest
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			boolean actual = false;
-			new Requirements(scope).requireThat(actual, "actual").isFalse();
+			new TestValidatorsImpl(scope).requireThat(actual, "Actual").isFalse();
 		}
 	}
 
@@ -44,29 +41,7 @@ public final class BooleanTest
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			boolean actual = true;
-			new Requirements(scope).requireThat(actual, "actual").isFalse();
-		}
-	}
-
-	@SuppressWarnings("deprecation")
-	@Test(expectedExceptions = IllegalArgumentException.class)
-	public void booleanIsNull_deprecation()
-	{
-		try (ApplicationScope scope = new TestApplicationScope(NONE))
-		{
-			boolean actual = true;
-			new Requirements(scope).requireThat(actual, "actual").isNull();
-		}
-	}
-
-	@SuppressWarnings("deprecation")
-	@Test
-	public void booleanIsNotNull_deprecation()
-	{
-		try (ApplicationScope scope = new TestApplicationScope(NONE))
-		{
-			boolean actual = true;
-			new Requirements(scope).requireThat(actual, "actual").isNotNull();
+			new TestValidatorsImpl(scope).requireThat(actual, "Actual").isFalse();
 		}
 	}
 
@@ -76,8 +51,8 @@ public final class BooleanTest
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			boolean actual = false;
-			new Requirements(scope).requireThat(actual, "actual").isEqualTo(true);
-			assert (false) : "Expected verifier to throw an exception";
+			new TestValidatorsImpl(scope).requireThat(actual, "Actual").isEqualTo(true);
+			assert (false) : "Expected validator to throw an exception";
 		}
 		catch (IllegalArgumentException e)
 		{
@@ -89,32 +64,30 @@ public final class BooleanTest
 	}
 
 	@Test
-	public void validateThatNullIsTrue()
+	public void multipleFailuresNullIsTrue()
 	{
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			Boolean actual = null;
-			List<String> expectedMessages = Collections.singletonList("actual may not be null");
-			List<ValidationFailure> actualFailures = new Requirements(scope).validateThat(actual, "actual").
-				isTrue().isEqualTo(5).getFailures();
-			List<String> actualMessages = actualFailures.stream().map(ValidationFailure::getMessage).
-				collect(Collectors.toList());
-			new Requirements(scope).requireThat(actualMessages, "actualMessages").isEqualTo(expectedMessages);
+			List<String> expectedMessages = List.of("\"Actual\" may not be null",
+				"\"Actual\" must be equal to 5");
+			List<String> actualMessages = new TestValidatorsImpl(scope).checkIf(actual, "Actual").
+				isTrue().isEqualTo(5).elseGetMessages();
+			new TestValidatorsImpl(scope).requireThat(actualMessages, "actualMessages").isEqualTo(expectedMessages);
 		}
 	}
 
 	@Test
-	public void validateThatNullIsFalse()
+	public void multipleFailuresNullIsFalse()
 	{
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			Boolean actual = null;
-			List<String> expectedMessages = Collections.singletonList("actual may not be null");
-			List<ValidationFailure> actualFailures = new Requirements(scope).validateThat(actual, "actual").
-				isFalse().isEqualTo(5).getFailures();
-			List<String> actualMessages = actualFailures.stream().map(ValidationFailure::getMessage).
-				collect(Collectors.toList());
-			new Requirements(scope).requireThat(actualMessages, "actualMessages").isEqualTo(expectedMessages);
+			List<String> expectedMessages = List.of("\"Actual\" may not be null",
+				"\"Actual\" must be equal to 5");
+			List<String> actualMessages = new TestValidatorsImpl(scope).checkIf(actual, "Actual").
+				isFalse().isEqualTo(5).elseGetMessages();
+			new TestValidatorsImpl(scope).requireThat(actualMessages, "actualMessages").isEqualTo(expectedMessages);
 		}
 	}
 }
