@@ -4,20 +4,17 @@
  */
 package com.github.cowwoc.requirements.test.java;
 
-import com.github.cowwoc.requirements.Requirements;
-import com.github.cowwoc.requirements.java.ValidationFailure;
 import com.github.cowwoc.requirements.java.internal.scope.ApplicationScope;
-import com.github.cowwoc.requirements.test.natives.internal.util.scope.TestApplicationScope;
+import com.github.cowwoc.requirements.test.TestValidatorsImpl;
+import com.github.cowwoc.requirements.test.scope.TestApplicationScope;
 import org.testng.annotations.Test;
 
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URL;
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import static com.github.cowwoc.requirements.natives.terminal.TerminalEncoding.NONE;
+import static com.github.cowwoc.requirements.java.terminal.TerminalEncoding.NONE;
 
 @SuppressWarnings("ConstantConditions")
 public final class UrlTest
@@ -27,8 +24,8 @@ public final class UrlTest
 	{
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
-			URL actual = new URL("http://host.com/");
-			new Requirements(scope).requireThat(actual, null);
+			URL actual = URI.create("http://host.com/").toURL();
+			new TestValidatorsImpl(scope).requireThat(actual, null);
 		}
 	}
 
@@ -37,8 +34,8 @@ public final class UrlTest
 	{
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
-			URL actual = new URL("http://host.com/");
-			new Requirements(scope).requireThat(actual, "");
+			URL actual = URI.create("http://host.com/").toURL();
+			new TestValidatorsImpl(scope).requireThat(actual, "");
 		}
 	}
 
@@ -48,7 +45,7 @@ public final class UrlTest
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			String actual = "http://host.com/index.html";
-			URL actualAsUrl = new Requirements(scope).requireThat(actual, "actual").asUrl().getActual();
+			URL actualAsUrl = new TestValidatorsImpl(scope).requireThat(actual, "Actual").asUrl().getValue();
 			assert (actualAsUrl.toString().equals(actual)) : "actualAsUrl: " + actualAsUrl + ", actual: " + actual;
 		}
 	}
@@ -59,48 +56,41 @@ public final class UrlTest
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			String actual = "http://host.com/index.html";
-			URI actualAsUri = new Requirements(scope).requireThat(actual, "actual").asUrl().asUri().
-				getActual();
+			URI actualAsUri = new TestValidatorsImpl(scope).requireThat(actual, "Actual").asUrl().asUri().
+				getValue();
 			assert (actualAsUri.toString().equals(actual)) : "actualAsUri: " + actualAsUri + ", actual: " + actual;
 		}
 	}
 
 	@Test
-	public void assertionsDisabled()
+	public void multipleFailuresAsUrlNull()
 	{
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
-			// Ensure that no exception is thrown if assertions are disabled
-			URL actual = null;
-			new Requirements(scope).withAssertionsDisabled().assertThat(r ->
-				r.requireThat(actual, "actual").isNotNull());
-		}
-	}
-
-	@Test(expectedExceptions = NullPointerException.class)
-	public void validateThatAsUrlNull()
-	{
-		try (ApplicationScope scope = new TestApplicationScope(NONE))
-		{
-			URL actual = null;
-			new Requirements(scope).validateThat(actual, "actual").asUri(null);
+			URI actual = null;
+			List<String> expectedMessages = List.of("\"Actual\" may not be null",
+				"""
+					"Actual" must be equal to "notEqual"\
+					""");
+			List<String> actualMessages = new TestValidatorsImpl(scope).checkIf(actual, "Actual").
+				isAbsolute().isEqualTo("notEqual").elseGetMessages();
+			new TestValidatorsImpl(scope).requireThat(actualMessages, "actualMessages").isEqualTo(expectedMessages);
 		}
 	}
 
 	@Test
-	public void validateThatNullAsUri()
+	public void multipleFailuresNullAsUri()
 	{
 		try (ApplicationScope scope = new TestApplicationScope(NONE))
 		{
 			URL actual = null;
-			List<String> expectedMessages = Collections.singletonList("""
-				actual must be a URI.
-				Actual: null""");
-			List<ValidationFailure> actualFailures = new Requirements(scope).validateThat(actual, "actual").
-				asUri().isEqualTo("notEqual").getFailures();
-			List<String> actualMessages = actualFailures.stream().map(ValidationFailure::getMessage).
-				collect(Collectors.toList());
-			new Requirements(scope).requireThat(actualMessages, "actualMessages").isEqualTo(expectedMessages);
+			List<String> expectedMessages = List.of("\"Actual\" may not be null",
+				"""
+					"Actual" must be equal to "notEqual"\
+					""");
+			List<String> actualMessages = new TestValidatorsImpl(scope).checkIf(actual, "Actual").
+				asUri().isEqualTo("notEqual").elseGetMessages();
+			new TestValidatorsImpl(scope).requireThat(actualMessages, "actualMessages").isEqualTo(expectedMessages);
 		}
 	}
 }
