@@ -1,114 +1,77 @@
 /*
- * Copyright (c) 2019 Gili Tzabari
+ * Copyright (c) 2014 Gili Tzabari
  * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
 package com.github.cowwoc.requirements.guava;
 
-import com.github.cowwoc.requirements.java.CollectionValidator;
-import com.github.cowwoc.requirements.java.SizeValidator;
-import com.github.cowwoc.requirements.java.extension.ExtensibleObjectValidator;
+import com.github.cowwoc.requirements.java.type.CollectionValidator;
+import com.github.cowwoc.requirements.java.type.PrimitiveUnsignedIntegerValidator;
+import com.github.cowwoc.requirements.java.type.part.ObjectPart;
+import com.github.cowwoc.requirements.java.type.part.Validator;
 import com.google.common.collect.Multimap;
 
 import java.util.Collection;
 import java.util.Map.Entry;
 import java.util.Set;
-import java.util.function.Consumer;
 
 /**
- * Validates the requirements of a Multimap.
+ * Validates the state of a Multimap, throwing an exception on the first failure.
  *
- * @param <K> the type of key in the multimap
- * @param <V> the type of value in the multimap
+ * @param <K> the type of keys in the {@code Multimap}
+ * @param <V> the type of values in the {@code Multimap}
+ * @param <T> the type of the {@code Multimap}
  */
-public interface MultimapValidator<K, V>
-	extends ExtensibleObjectValidator<MultimapValidator<K, V>, Multimap<K, V>>
+public interface MultimapValidator<K, V, T extends Multimap<K, V>> extends
+	Validator<MultimapValidator<K, V, T>>,
+	ObjectPart<MultimapValidator<K, V, T>, T>
 {
 	/**
-	 * Returns a validator over {@link Multimap#keySet()}.
+	 * Ensures that the Multimap is empty.
 	 *
-	 * @return a validator over {@link Multimap#keySet()}
+	 * @return this
+	 * @throws NullPointerException     if the value is null
+	 * @throws IllegalArgumentException if the Multimap is not empty
 	 */
-	CollectionValidator<Set<K>, K> keySet();
+	MultimapValidator<K, V, T> isEmpty();
 
 	/**
-	 * Validates nested requirements. This mechanism can be used to
-	 * <a href="https://github.com/cowwoc/requirements.java/wiki/Features.md#grouping-nested-requirements">
-	 * group related requirements</a>.
+	 * Ensures that the Multimap is not empty.
 	 *
-	 * @param consumer validates the {@link Multimap#keySet()}
-	 * @return the updated validator
+	 * @return this
+	 * @throws NullPointerException     if the value is null
+	 * @throws IllegalArgumentException if the Multimap is empty
 	 */
-	@SuppressWarnings("LongLine")
-	MultimapValidator<K, V> keySet(Consumer<CollectionValidator<Set<K>, K>> consumer);
+	MultimapValidator<K, V, T> isNotEmpty();
 
 	/**
-	 * Returns a validator over {@link Multimap#values()}.
+	 * Returns a validator over the Multimap's {@link Multimap#keySet() keys}.
 	 *
-	 * @return a validator over {@link Multimap#values()}
+	 * @return a validator over the Multimap's {@link Multimap#keySet() keys}
+	 * @throws NullPointerException if the value is null
 	 */
-	CollectionValidator<Collection<V>, V> values();
+	CollectionValidator<K, Set<K>> keySet();
 
 	/**
-	 * Validates nested requirements. This mechanism can be used to
-	 * <a href="https://github.com/cowwoc/requirements.java/wiki/Features.md#grouping-nested-requirements">
-	 * group related requirements</a>.
+	 * Returns a validator over the Multimap's {@link Multimap#values() values}.
 	 *
-	 * @param consumer validates the {@link Multimap#values()}
-	 * @return the updated validator
+	 * @return a validator over the Multimap's {@link Multimap#values() values}
+	 * @throws NullPointerException if the value is null
 	 */
-	@SuppressWarnings("LongLine")
-	MultimapValidator<K, V> values(Consumer<CollectionValidator<Collection<V>, V>> consumer);
+	CollectionValidator<V, Collection<V>> values();
 
 	/**
-	 * Reutrns a validator over {@link Multimap#entries()}.
+	 * Returns a validator over the Multimap's {@link Multimap#entries() entries}.
 	 *
-	 * @return a validator over {@link Multimap#entries()}
+	 * @return a validator over the Multimap's {@link Multimap#entries() entries}
+	 * @throws NullPointerException if the value is null
 	 */
-	CollectionValidator<Collection<Entry<K, V>>, Entry<K, V>> entries();
+	CollectionValidator<Entry<K, V>, Collection<Entry<K, V>>> entries();
 
 	/**
-	 * Validates nested requirements. This mechanism can be used to
-	 * <a href="https://github.com/cowwoc/requirements.java/wiki/Features.md#grouping-nested-requirements">
-	 * group related requirements</a>.
+	 * Returns a validator over the Multimap's {@link Multimap#size() size}.
 	 *
-	 * @param consumer validates the {@link Multimap#entries()}
-	 * @return the updated validator
+	 * @return a validator over the Multimap's {@link Multimap#size() size}.
+	 * @throws NullPointerException if the value is null
 	 */
-	@SuppressWarnings("LongLine")
-	MultimapValidator<K, V> entries(Consumer<CollectionValidator<Collection<Entry<K, V>>,
-		Entry<K, V>>> consumer);
-
-	/**
-	 * Ensures that the actual value is empty.
-	 *
-	 * @return the updated validator
-	 * @throws IllegalArgumentException if actual value is not empty
-	 */
-	MultimapValidator<K, V> isEmpty();
-
-	/**
-	 * Ensures that the actual value is not empty.
-	 *
-	 * @return the updated validator
-	 * @throws IllegalArgumentException if actual value is empty
-	 */
-	MultimapValidator<K, V> isNotEmpty();
-
-	/**
-	 * Returns a validator over {@link Multimap#size()}.
-	 *
-	 * @return a validator over {@link Multimap#size()}
-	 */
-	SizeValidator size();
-
-	/**
-	 * Validates nested requirements. This mechanism can be used to
-	 * <a href="https://github.com/cowwoc/requirements.java/wiki/Features.md#grouping-nested-requirements">
-	 * group related requirements</a>.
-	 *
-	 * @param consumer validates the multimap's size
-	 * @return the updated validator
-	 */
-	@SuppressWarnings("LongLine")
-	MultimapValidator<K, V> size(Consumer<SizeValidator> consumer);
+	PrimitiveUnsignedIntegerValidator size();
 }
