@@ -1,6 +1,7 @@
 package com.github.cowwoc.requirements.java.internal.util;
 
-import com.github.cowwoc.requirements.java.internal.util.WrappedCheckedException.CallableWithoutReturnValue;
+import com.github.cowwoc.requirements.java.WrappedCheckedException;
+import com.github.cowwoc.requirements.java.WrappedCheckedException.Task;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.locks.StampedLock;
@@ -25,14 +26,18 @@ public final class ReentrantStampedLock
 	{
 	}
 
+	private static void doNotUnlock()
+	{
+	}
+
 	/**
-	 * Acquires an optimistic read-lock and runs a task. If an optimistic read-lock cannot be acquired,
-	 * invokes {@link #read(Callable) readLock(task)} instead.
+	 * Acquires an optimistic read-lock and runs a task. If an optimistic read-lock cannot be acquired, invokes
+	 * {@link #read(Callable) readLock(task)} instead.
 	 * <p>
-	 * {@code task} is guaranteed to be invoked <i>at least</i> once, but must be safe to invoke multiple
-	 * times as well. The return value must correspond to a local copy of the fields being read as there is
-	 * no guarantee that state won't change between the time the lock is released and the time that the value
-	 * is returned.
+	 * {@code task} is guaranteed to be invoked <i>at least</i> once, but must be safe to invoke multiple times
+	 * as well. The return value must correspond to a local copy of the fields being read as there is no
+	 * guarantee that state won't change between the time the lock is released and the time that the value is
+	 * returned.
 	 *
 	 * @param <V>  the type of value returned by the task
 	 * @param task the task to run while holding the lock
@@ -64,19 +69,19 @@ public final class ReentrantStampedLock
 	}
 
 	/**
-	 * Acquires an optimistic read-lock and runs a task. If an optimistic read-lock cannot be acquired,
-	 * invokes {@link #read(CallableWithoutReturnValue) readLock(task)} instead.
+	 * Acquires an optimistic read-lock and runs a task. If an optimistic read-lock cannot be acquired, invokes
+	 * {@link #read(Task) readLock(task)} instead.
 	 * <p>
-	 * {@code task} is guaranteed to be invoked <i>at least</i> once, but must be safe to invoke multiple
-	 * times as well. The return value must correspond to a local copy of the fields being read as there is
-	 * no guarantee that state won't change between the time the lock is released and the time that the value
-	 * is returned.
+	 * {@code task} is guaranteed to be invoked <i>at least</i> once, but must be safe to invoke multiple times
+	 * as well. The return value must correspond to a local copy of the fields being read as there is no
+	 * guarantee that state won't change between the time the lock is released and the time that the value is
+	 * returned.
 	 *
 	 * @param task the task to run while holding the lock
 	 * @throws NullPointerException    if {@code task} is null
 	 * @throws WrappedCheckedException if any checked exceptions are thrown
 	 */
-	public void optimisticRead(CallableWithoutReturnValue task)
+	public void optimisticRead(Task task)
 	{
 		optimisticRead(() ->
 		{
@@ -86,13 +91,13 @@ public final class ReentrantStampedLock
 	}
 
 	/**
-	 * Acquires an optimistic read-lock and runs a task. If an optimistic read-lock cannot be acquired,
-	 * invokes {@link #read(Callable) readLock(task)} instead.
+	 * Acquires an optimistic read-lock and runs a task. If an optimistic read-lock cannot be acquired, invokes
+	 * {@link #read(Callable) readLock(task)} instead.
 	 * <p>
-	 * {@code task} is guaranteed to be invoked <i>at least</i> once, but must be safe to invoke multiple
-	 * times as well. The return value must correspond to a local copy of the fields being read as there is
-	 * no guarantee that state won't change between the time the lock is released and the time that the value
-	 * is returned.
+	 * {@code task} is guaranteed to be invoked <i>at least</i> once, but must be safe to invoke multiple times
+	 * as well. The return value must correspond to a local copy of the fields being read as there is no
+	 * guarantee that state won't change between the time the lock is released and the time that the value is
+	 * returned.
 	 *
 	 * @param <V>   the type of value returned by the task
 	 * @param task  the task to run while holding the lock
@@ -113,8 +118,8 @@ public final class ReentrantStampedLock
 	}
 
 	/**
-	 * Acquires a read-lock, unless the caller already holds a read or write-lock. If the caller already
-	 * holds a lock, no lock is acquired or released.
+	 * Acquires a read-lock, unless the caller already holds a read or write-lock. If the caller already holds a
+	 * lock, no lock is acquired or released.
 	 *
 	 * @return a read-lock as a resource
 	 */
@@ -164,7 +169,7 @@ public final class ReentrantStampedLock
 	 * @throws NullPointerException    if {@code task} is null
 	 * @throws WrappedCheckedException if {@code task} throws a checked exception
 	 */
-	public void read(CallableWithoutReturnValue task)
+	public void read(Task task)
 	{
 		read(() ->
 		{
@@ -174,13 +179,13 @@ public final class ReentrantStampedLock
 	}
 
 	/**
-	 * Acquires a write-lock, unless the caller already holds one. If the caller already holds a lock, no
-	 * lock is acquired or released.
+	 * Acquires a write-lock, unless the caller already holds one. If the caller already holds a lock, no lock
+	 * is acquired or released.
 	 * <p>
 	 * If the caller holds a read-lock, an attempt is made to convert it into a write-lock. Conversions are not
-	 * guaranteed to be atomic; consequently, there is no guarantee that state won't change between the time
-	 * the read-lock is released and the write-lock is acquired. Callers must repeat any state checks to
-	 * ensure that they still hold.
+	 * guaranteed to be atomic; consequently, there is no guarantee that state won't change between the time the
+	 * read-lock is released and the write-lock is acquired. Callers must repeat any state checks to ensure that
+	 * they still hold.
 	 *
 	 * @return a write-lock as a resource
 	 */
@@ -235,9 +240,9 @@ public final class ReentrantStampedLock
 	 * released.
 	 * <p>
 	 * If the caller holds a read-lock, an attempt is made to convert it into a write-lock. Conversions are not
-	 * guaranteed to be atomic; consequently, there is no guarantee that state won't change between the time
-	 * the read-lock is released and the write-lock is acquired. Callers must repeat any state checks to
-	 * ensure that they still hold.
+	 * guaranteed to be atomic; consequently, there is no guarantee that state won't change between the time the
+	 * read-lock is released and the write-lock is acquired. Callers must repeat any state checks to ensure that
+	 * they still hold.
 	 *
 	 * @param <V>  the type of value returned by the task
 	 * @param task the task to run while holding the lock
@@ -258,15 +263,15 @@ public final class ReentrantStampedLock
 	 * released.
 	 * <p>
 	 * If the caller holds a read-lock, an attempt is made to convert it into a write-lock. Conversions are not
-	 * guaranteed to be atomic; consequently, there is no guarantee that state won't change between the time
-	 * the read-lock is released and the write-lock is acquired. Callers must repeat any state checks to
-	 * ensure that they still hold.
+	 * guaranteed to be atomic; consequently, there is no guarantee that state won't change between the time the
+	 * read-lock is released and the write-lock is acquired. Callers must repeat any state checks to ensure that
+	 * they still hold.
 	 *
 	 * @param task the task to run while holding the lock
 	 * @throws NullPointerException    if {@code task} is null
 	 * @throws WrappedCheckedException if {@code task} throws a checked exception
 	 */
-	public void write(CallableWithoutReturnValue task)
+	public void write(Task task)
 	{
 		write(() ->
 		{
@@ -294,9 +299,5 @@ public final class ReentrantStampedLock
 		{
 			throw WrappedCheckedException.wrap(e);
 		}
-	}
-
-	private static void doNotUnlock()
-	{
 	}
 }
